refactor(repo C11b): extract embedding storage to engine/repository/embeddings.ts
Move getEmbedding plus the embedding-storage layer (getHashesForEmbedding,
clearAllEmbeddings, insertEmbedding, bulkInsertEmbeddings, and their
WeakMap-cached prepared-statement and transaction wrappers) out of
engine/repository.ts. vec.ts continues calling getEmbedding via the
parent re-export; remaining repository.ts callers import directly.

Build: clean. Tests: 970/970 pass.
